feat(gui): 重构管理员界面并添加新功能
- 重新设计了管理员界面的布局和样式,增加了更多视觉元素
- 添加了管理员信息面板,展示当前登录管理员的详细信息
- 实现了带图标的自定义选项卡面板
- 增加了通用的样式方法,提升了代码复用性
- 优化了数据库配置,统一了JPA设置
- 调整了多个模型类的ID类型,从Integer改为Long

+++ backend_pj/src/main/java/com/carpj/model/Material.java
@@ -12,7 +12,7 @@ public class Material {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "material_id")
-    private Integer id;
+    private Long id;
     
     @Column(nullable = false, length = 100)
     private String name;
@@ -24,10 +24,10 @@ public class Material {
     private BigDecimal unitPrice;
     
     @Column(name = "current_stock")
-    private Integer currentStock = 0;
+    private Long currentStock = 0L;
     
     @Column(name = "min_stock_level")
-    private Integer minStockLevel = 5;
+    private Long minStockLevel = 5L;
     
     @Column(length = 100)
     private String supplier;
